+++ engines/sci/graphics/drivers/hercules.cpp
@@ -1,0 +1,196 @@
+/* ScummVM - Graphic Adventure Engine
+ *
+ * ScummVM is the legal property of its developers, whose names
+ * are too numerous to list here. Please refer to the COPYRIGHT
+ * file distributed with this source distribution.
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ *
+ */
+
+
+#include "common/file.h"
+#include "common/system.h"
+#include "graphics/cursorman.h"
+#include "sci/graphics/drivers/gfxdriver_intern.h"
+
+namespace Sci {
+
+class SCI0_HerculesDriver final : public SCI0_DOSPreVGADriver {
+public:
+	SCI0_HerculesDriver(uint32 monochromeColor, bool rgbRendering, bool cropImage);
+	~SCI0_HerculesDriver() override;
+	void copyRectToScreen(const byte *src, int srcX, int srcY, int pitch, int destX, int destY, int w, int h, const PaletteMod*, const byte*) override;
+	void replaceCursor(const void *cursor, uint w, uint h, int hotspotX, int hotspotY, uint32 keycolor) override;
+	Common::Point getMousePos() const override;
+	void setMousePos(const Common::Point &pos) const override;
+	void setShakePos(int shakeXOffset, int shakeYOffset) const override;
+	void clearRect(const Common::Rect &r) const override;
+	Common::Point getRealCoords(Common::Point &pos) const override;
+	static bool validateMode(Common::Platform p) { return (p == Common::kPlatformDOS) && checkDriver(&_driverFile, 1); }
+private:
+	void setupRenderProc() override;
+	const uint16 _centerX;
+	const uint16 _centerY;
+	byte _monochromePalette[6];
+	const byte *_monochromePatterns;
+	typedef void (*LineProc)(byte*&, const byte*, int, int, int, const byte*, const byte*);
+	LineProc _renderLine;
+	static const char *_driverFile;
+};
+
+SCI0_HerculesDriver::SCI0_HerculesDriver(uint32 monochromeColor, bool rgbRendering, bool cropImage) : SCI0_DOSPreVGADriver(2, cropImage ? 640 : 720, cropImage ? 300 : 350, rgbRendering),
+	_centerX(cropImage ? 0 : 40), _centerY(cropImage ? 0 : 25), _monochromePatterns(nullptr), _renderLine(nullptr) {
+	_monochromePalette[0] = _monochromePalette[1] = _monochromePalette[2] = 0;
+	_monochromePalette[3] = (monochromeColor >> 16) & 0xff;
+	_monochromePalette[4] = (monochromeColor >> 8) & 0xff;
+	_monochromePalette[5] = monochromeColor & 0xff;
+	assignPalette(_monochromePalette);
+	bool unused = false;
+
+	if (!(_monochromePatterns = SciGfxDrvInternal::monochrInit(_driverFile, unused)))
+		GFXDRV_ERR_OPEN(_driverFile);
+}
+
+SCI0_HerculesDriver::~SCI0_HerculesDriver() {
+	delete[] _monochromePatterns;
+}
+
+void SCI0_HerculesDriver::copyRectToScreen(const byte *src, int srcX, int srcY, int pitch, int destX, int destY, int w, int h, const PaletteMod*, const byte*) {
+	GFXDRV_ASSERT_READY;
+
+	byte *dst = _compositeBuffer;
+	byte sw = destY & 1;
+	src += (srcY * pitch + srcX);
+	destY = (destY & ~1) * 3 / 2 + (destY & 1);
+	int ty = destY & 7;
+	int rh = 0;
+
+	for (int i = 0; i < h; ++i) {
+		const byte *src2 = src;
+		_renderLine(dst, src2, w, srcX & 3, ty, _monochromePatterns, _internalPalette);
+		ty = (ty + 1) & 7;
+		++rh;
+
+		if (sw & 1)
+			sw ^= 2;
+
+		if (sw != 3) {
+			src += pitch;
+			sw ^= 1;
+		} else {
+			--i;
+		}
+	}
+
+	g_system->copyRectToScreen(_compositeBuffer, (w << 1) * _pixelSize, (destX << 1) + _centerX, destY + _centerY, w << 1, rh);
+}
+
+void SCI0_HerculesDriver::replaceCursor(const void *cursor, uint w, uint h, int hotspotX, int hotspotY, uint32 keycolor) {
+	GFXDRV_ASSERT_READY;
+	// Instead of implementing the original cursor rendering code, we rely on the 8 bit cursor that
+	// has already been generated by the engine. We simply convert the colors as needed and scale the cursor...
+	assert(keycolor == 1);
+	keycolor = 0x0f;
+	int alt = 0;
+	const byte *s = reinterpret_cast<const byte *>(cursor);
+	byte *d = _compositeBuffer;
+
+	for (uint i = 0; i < h; ++i) {
+		for (uint ii = 0; ii < (w << 1); ++ii) {
+			*d++ = *s ? (*s ^ 0x0e) : 0;
+			if (ii & 1)
+				++s;
+		}
+		if (i & 1) {
+			alt ^= 1;
+			if (alt) {
+				s -= w;
+				--i;
+			}
+		}
+	}
+
+	CursorMan.replaceCursor(_compositeBuffer, w << 1, (h & ~1) * 3 / 2 + (h & 1), hotspotX << 1, (hotspotY & ~1) * 3 / 2 + (hotspotY & 1), keycolor);
+}
+
+Common::Point SCI0_HerculesDriver::getMousePos() const {
+	Common::Point res = GfxDriver::getMousePos();
+	res.x = CLIP<int>(res.x - _centerX, 0, 639) >> 1;
+	res.y = (CLIP<int>(res.y - _centerY, 0, 299) * 2 + 1) / 3;
+	return res;
+}
+
+void SCI0_HerculesDriver::setMousePos(const Common::Point &pos) const {
+	g_system->warpMouse((pos.x << 1) + _centerX, (pos.y & ~1) * 3 / 2 + (pos.y & 1) + _centerY);
+}
+
+void SCI0_HerculesDriver::setShakePos(int shakeXOffset, int shakeYOffset) const {
+	g_system->setShakePos(shakeXOffset << 1, (shakeYOffset & ~1) * 3 / 2 + (shakeYOffset & 1));
+}
+
+void SCI0_HerculesDriver::clearRect(const Common::Rect &r) const {
+	Common::Rect r2((r.left << 1) + _centerX, (r.top & ~1) * 3 / 2 + (r.top & 1) + _centerY, (r.right << 1) + 40, (r.bottom & ~1) * 3 / 2 + (r.bottom & 1) + _centerY);
+	GfxDriver::clearRect(r2);
+}
+
+Common::Point SCI0_HerculesDriver::getRealCoords(Common::Point &pos) const {
+	return Common::Point((pos.x << 1) + _centerX, (pos.y & ~1) * 3 / 2 + (pos.y & 1) + _centerY);
+}
+
+template <typename T> void herculesRenderLine(byte *&dst, const byte *src, int w, int tx, int ty, const byte *patterns, const byte *pal) {
+	T *d = reinterpret_cast<T*>(dst);
+	const T *p = reinterpret_cast<const T*>(pal);
+
+	for (int i = 0; i < w; ++i) {
+		byte pt = patterns[((*src++ & 0x0f) << 3) + ty] >> (6 - (tx << 1));
+		if (sizeof(T) == 1) {
+			*d++ = (pt >> 1) & 1;
+			*d++ = pt & 1;
+		} else {
+			*d++ = p[(pt >> 1) & 1];
+			*d++ = p[pt & 1];
+		}
+		tx = (tx + 1) & 3;
+	}
+
+	dst = reinterpret_cast<byte*>(d);
+}
+
+void SCI0_HerculesDriver::setupRenderProc() {
+	static const LineProc lineProcs[] = {
+		&herculesRenderLine<byte>,
+		&herculesRenderLine<uint16>,
+		&herculesRenderLine<uint32>
+	};
+
+	assert((_pixelSize >> 1) < ARRAYSIZE(lineProcs));
+	_renderLine = lineProcs[_pixelSize >> 1];
+}
+
+const char *SCI0_HerculesDriver::_driverFile = "HERCMONO.DRV";
+
+SCI_GFXDRV_VALIDATE_IMPL(SCI0_Hercules)
+
+GfxDriver *SCI0_HerculesDriver_create(int rgbRendering, ...) {
+	static const uint32 monochromeColors[] = { 0xffbf66, 0x66ff66, 0xffffff };
+	va_list args;
+	va_start(args, rgbRendering);
+	int config = CLIP<int>(va_arg(args, int), 0, ARRAYSIZE(monochromeColors));
+	va_end(args);
+
+	return new SCI0_HerculesDriver(monochromeColors[config], rgbRendering != 0, false);
+}
+
+} // End of namespace Sci
