+++ engines/sci/module.mk
@@ -50,7 +50,6 @@ MODULE_OBJS := \
 	graphics/cursor.o \
 	graphics/fontkorean.o \
 	graphics/fontsjis.o \
-	graphics/gfxdrivers.o \
 	graphics/macfont.o \
 	graphics/maciconbar.o \
 	graphics/menu.o \
@@ -66,6 +65,20 @@ MODULE_OBJS := \
 	graphics/text16.o \
 	graphics/transitions.o \
 	graphics/view.o \
+	graphics/drivers/cga.o \
+	graphics/drivers/cgabw.o \
+	graphics/drivers/common.o \
+	graphics/drivers/default.o \
+	graphics/drivers/ega.o \
+	graphics/drivers/hercules.o \
+	graphics/drivers/init.o \
+	graphics/drivers/pc98_8col_sci0.o \
+	graphics/drivers/pc98_8col_sci1.o \
+	graphics/drivers/pc98_16col.o \
+	graphics/drivers/upscaled.o \
+	graphics/drivers/vgagrey.o \
+	graphics/drivers/win16col.o \
+	graphics/drivers/win256col.o \
 	parser/grammar.o \
 	parser/said.o \
 	parser/vocabulary.o \
